+++ astroquery/ipac/irsa/tests/setup_package.py
@@ -5,6 +5,9 @@ import os
 
 
 def get_package_data():
-    paths = [os.path.join('data', '*.xml'),
-             ]
+    paths = [
+        os.path.join('data', '*.xml'),
+        os.path.join('data', '*.html'),
+        os.path.join('data', '*.tbl'),
+    ]
     return {'astroquery.ipac.irsa.tests': paths}